GET Method reading info from request
